fix: access shared trashbin objects
fix: access shared trashbin objects



[skip ci]

+++ apps/dav/lib/CalDAV/Trashbin/DeletedCalendarObject.php
@@ -10,6 +10,7 @@ namespace OCA\DAV\CalDAV\Trashbin;
 
 use OCA\DAV\CalDAV\CalDavBackend;
 use OCA\DAV\CalDAV\IRestorable;
+use OCA\DAV\DAV\Sharing\Backend;
 use Sabre\CalDAV\ICalendarObject;
 use Sabre\DAV\Exception\Forbidden;
 use Sabre\DAVACL\ACLTrait;
@@ -28,6 +29,9 @@ class DeletedCalendarObject implements IACL, ICalendarObject, IRestorable {
 	}
 
 	public function delete() {
+		if (!$this->canModify()) {
+			throw new Forbidden('Read-only sharees cannot permanently delete trashbin entries');
+		}
 		$this->calDavBackend->deleteCalendarObject(
 			$this->objectData['calendarid'],
 			$this->objectData['uri'],
@@ -74,6 +78,9 @@ class DeletedCalendarObject implements IACL, ICalendarObject, IRestorable {
 	}
 
 	public function restore(): void {
+		if (!$this->canModify()) {
+			throw new Forbidden('Read-only sharees cannot restore trashbin entries');
+		}
 		$this->calDavBackend->restoreCalendarObject($this->objectData);
 	}
 
@@ -86,19 +93,14 @@ class DeletedCalendarObject implements IACL, ICalendarObject, IRestorable {
 	}
 
 	public function getACL(): array {
-		return [
+		$acl = [
 			[
 				'privilege' => '{DAV:}read', // For queries
 				'principal' => $this->getOwner(),
 				'protected' => true,
 			],
 			[
-				'privilege' => '{DAV:}unbind', // For moving and deletion
-				'principal' => $this->getOwner(),
-				'protected' => true,
-			],
-			[
-				'privilege' => '{DAV:}all',
+				'privilege' => '{DAV:}read',
 				'principal' => $this->getOwner() . '/calendar-proxy-write',
 				'protected' => true,
 			],
@@ -108,6 +110,23 @@ class DeletedCalendarObject implements IACL, ICalendarObject, IRestorable {
 				'protected' => true,
 			],
 		];
+
+		if ($this->canModify()) {
+			$acl[] = [
+				'privilege' => '{DAV:}unbind',
+				'principal' => $this->getOwner(),
+				'protected' => true,
+			];
+
+			$acl[] = [
+				'privilege' => '{DAV:}unbind',
+				'principal' => $this->getOwner() . '/calendar-proxy-write',
+				'protected' => true,
+			];
+
+		}
+
+		return $acl;
 	}
 
 	public function getOwner() {
